requirements.txt
# Core ML
lightgbm>=4.3.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0

# Data storage
duckdb>=0.10.0
pyarrow>=15.0.0

# Experiment tracking & model registry
mlflow>=2.12.0

# Orchestration
prefect>=2.19.0

# API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.3
python-multipart>=0.0.9

# Monitoring
evidently>=0.4.22

# Agent / LLM tools
openai>=1.23.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
loguru>=0.7.2
httpx>=0.27.0
